escape 0.2
