+++ client/client_urls.py
@@ -9,9 +9,9 @@ urlpatterns = [
     path('profile', views.client_profile, name='client_profile'),
 
     # Team Rome - Edit Forms URLs
-    path('edit', views.client_edit_main, name='client_edit'),
-    path('edit/skills', views.client_edit_skills, name='client_edit_skills'),
-    path('edit/photo', views.client_edit_photo, name='client_edit_photo'),
+    path('edit', views.ClientEditMain.as_view(), name='client_edit'),
+    path('edit/skills', views.ClientEditSkills.as_view(), name='client_edit_skills'),
+    path('edit/photo', views.ClientEditPhoto.as_view(), name='client_edit_photo'),
     path('edit/cv', views.client_edit_cv, name='client_edit_cv'),
     path('edit/education', views.client_edit_education, name='client_edit_education'),
     path('edit/experience', views.client_edit_experience, name='client_edit_experience'),
